+++ NEWS.md
@@ -3,6 +3,7 @@
 - Add `layer` to `perplexity()` for `textmodel_doc2vec` models.
 - Save document lengths as `ntoken` in trained `textmodel_doc2vec` models.
 - Update `as.textmode_doc2vec()` to save output layer weights.
+- Update tests for **quanteda** v4.4.0.
 
 ## Changes in v0.6.1